requirements.txt
# PersonalAgent Dependencies
# Core Google Calendar Integration
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0

# Environment and Configuration
python-dotenv==1.1.1
python-dateutil==2.9.0.post0

# Security and Encryption
cryptography==44.0.0

# HTTP and Web
requests==2.32.4
httplib2==0.22.0

# Optional AI/LLM Dependencies
# Uncomment based on your preferred AI provider:

# OpenAI GPT Models
# openai>=1.0.0

# Google Gemini
# google-generativeai>=0.3.0

# Anthropic Claude
# anthropic>=0.3.0

# Core Dependencies (automatically installed with above)
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
google-api-core==2.25.1
googleapis-common-protos==1.70.0
idna==3.10
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
six==1.17.0
uritemplate==4.1.1
urllib3==2.2.4
